+++ package.json
@@ -11,6 +11,7 @@
   "scripts": {
     "start": "npm run build && node ./dist/main.js",
     "start:dev": "nodemon",
+    "start:compile": "tsx src/main-rest.ts",
     "build": "npm run clean && npm run compile",
     "lint": "eslint src/ --ext .ts",
     "compile": "tsc -p tsconfig.json",
@@ -38,6 +39,7 @@
     "rimraf": "5.0.1",
     "ts-node": "10.9.2",
     "tslib": "2.7.0",
+    "tsx": "4.19.2",
     "typescript": "5.4.5"
   },
   "type": "module",
