+++ src/shared/const/component.ts
