+++ src/shared/const/index.ts
@@ -1,0 +1,2 @@
+export { CityCoordinates } from './city-coordinates.js';
+export { Component } from './component.js';
