+++ src/shared/enum/index.ts
@@ -1,7 +1,6 @@
-export { CityName, CityCoordinates } from './city-information.js';
+export { CityName } from './city-information.js';
 export { HousingType } from './housing-type.js';
 export { Goods } from './goods.js';
 export { TypeUser } from './type-user.js';
-export { Component } from './component.js';
 export { ValueMockCoordinates } from './value-mock-coordinates.js';
 export { SortType } from './sort-type.js';
