+++ src/shared/modules/comment/comment-container.ts
@@ -1,9 +1,9 @@
 import { Container } from 'inversify';
 import { types } from '@typegoose/typegoose';
 import { CommentService } from '../../interface/index.js';
-import { Component } from '../../enum/index.js';
 import { DefaultCommentService } from './default-comment-service.js';
 import { CommentEntity, CommentModel } from './comment-entity.js';
+import { Component } from '../../const/index.js';
 
 export function createCommentContainer() {
   const commentContainer = new Container();
