+++ src/shared/modules/offer/offer-container.ts
@@ -1,9 +1,9 @@
 import { Container } from 'inversify';
 import { types } from '@typegoose/typegoose';
 import { OfferService } from '../../interface/index.js';
-import { Component } from '../../enum/index.js';
 import { OfferEntity, OfferModel } from '../offer/index.js';
 import { DefaultOfferService } from './default-offer-service.js';
+import { Component } from '../../const/index.js';
 
 export function createOfferContainer() {
   const offerContainer = new Container();
